Fixed an issue on the now method from the celery app
There was an issue on the celery app class. The now method was used to get the current time and date as datetime for the configured timezone. But the value was returned with datetime.utcnow() replacing the timezone. But the new datetime was neved recalculated, which caused that the datetime object had the utc datime but with the configured timezone, making it an aware datetime. The fix for this was to recalculate the the datetime with timezone.fromutc()